pkgfile module submission for summit
pkgfile module will allow users to add/delete/rollback any content within the roxie package file. This is an alternative for packagemap because packagemap is run on command line and for one to incorporate commands into ecl, it takes into different dimension, hence this module provides a straight forward ECL code to build package file. All information is stored in flat file on thor, if users want to build a package function pkgfile.RoxiePackage() will convert the flat to xml and also provides a function to despray the xml to any desired location.